+++ include/medium/2d/acoustic/isotropic/mass_matrix.tpp
