+++ include/medium/2d/acoustic/isotropic/properties.hpp
