+++ include/medium/2d/acoustic/isotropic/stress.hpp
