+++ include/medium/2d/acoustic/isotropic/stress_integrand.tpp
