+++ include/medium/2d/acoustic/isotropic/wavefield.hpp
