+++ include/medium/2d/elastic/anisotropic/frechet_derivative.hpp
