+++ include/medium/2d/elastic/anisotropic/mass_matrix.tpp
