+++ include/medium/2d/elastic/anisotropic/properties.hpp
