+++ include/medium/2d/elastic/anisotropic/stress.hpp
