+++ include/medium/2d/elastic/anisotropic/stress_integrand.hpp
