+++ include/medium/2d/elastic/isotropic/mass_matrix.hpp
