+++ include/medium/2d/elastic/isotropic/properties.hpp
