+++ include/medium/2d/elastic/isotropic/stress.hpp
