+++ include/medium/2d/elastic/isotropic/stress_integrand.tpp
