+++ include/medium/material.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "acoustic_isotropic2d/properties.hpp"
-#include "elastic_anisotropic2d/properties.hpp"
-#include "elastic_isotropic2d/properties.hpp"
+#include "2d/acoustic/isotropic/properties.hpp"
+#include "2d/elastic/anisotropic/properties.hpp"
+#include "2d/elastic/isotropic/properties.hpp"
 #include "enumerations/specfem_enums.hpp"
 #include "point/properties.hpp"
 #include "properties.hpp"
